Update display when answer is changed.
When the user first submits the answers, the checkmark/cross icons,
tip popups, and a message are shown to the user.

If the user changes some of the answers, the checkmarks, popups, and message
should automatically update. This is done by automatically resubmitting the answer.
Auto submission is only enabled if number of attempts is not limited (max_attempts = 0),