Step Builder: Lock and unlock navigation when appropriate.
When navigation is locked, it is not possible to click arrow buttons to
navigate between existing units. This should be the case during
completion of an attempt.

When navigation is unlocked, users can click arrow buttons to navigate
between existing units. This should be the case when user has completed
an attempt, i.e., when they are reviewing their grade for the last
attempt.