Patch wsgi handler to ignore broken pipe errors.
We would often see 'Broken pipe' errors pulluting the logs when running tests.
Broken pipe errors occur when browser aborts connection. It can be because
the browser was navigated to a different url before the current page finished
loading, but it could also be because it started downloading an image, but decided
that the image didn't change after loading part of it, and aborts the connection
to use the version from the cache.

The broken pipe errors are benign and are actually suppressed by default in Django 1.8.