Enable freeform answer submit when feedback is hidden.
When problem builder is configured to hide feedback upon revisiting a question
(pb_hide_feedback_if_attempts_remain option is enabled), when the user returns
to a problem builder block which contains only a freeform question,
the user is not able to click "Submit" without changing the answer text first.

This patch fixes that by keeping submit button enabled if
pb_hide_feedback_if_attempts_remain is turned on.

MCKIN-3890