Introduce Answer.course_key and deprecate course_id.
This introduces a new 255-char course_key field to the Answer model.
The old 50-char course_id field is still present, but is deprecated and
will be removed in next release.

We cannot simply extend the existing course_id field because we have
some large problem_builder_answer tables in production and the migration
to extend the column would lock the table causing issues in production.

The code is updated so that it uses the new course_key column, but falls
back to course_id.
In next release, a data migration to copy course_id data to course_key
will be provided and the course_id column dropped.