consolidate all test dependencies into test_requirements.txt
using the base.txt requirement files from xblock-sdk instead of
duplicating them is a workaround because the setup.py from xblock-sdk
does not install all dependencies (djpyfs is missing).

using the test.txt requirement file is not a good idea: there is no
reason why all test dependecies from xblock-sdk are needed to run the
tests in problem-builder. They actually conflict because of the
selenium version.

after xblock-sdk 0.1.4 the dependencies were reworked entirely to
comply with OEP-0018 and they can no longer be used as-is: it triggers
an error.