chore: disable pylint warning consider-using-f-string
Problem Builder code contains a lot of statements like this:

    _("[{status}]").format(status=status)

which appear that way because we use gettext to provide a translated
string for "[{status}]", which we then format by filling in the value
for status. Using an f-string here would break the translations, since
{status} would be filled in before the translation is fetched.

We can revisit this issue at a later date if there's a better way to
deal with this.